fix: adapt to LabelValue=i32 -- dereference iter_fixed_size, cast usize to i32
- iter_fixed_size now yields [&i32; N], use & in destructuring patterns
- builder.add with usize args needs explicit as i32 cast
- .into() on usize/integer literals replaced with as i32 or _i32 suffix

+++ featomic/src/calculator.rs
@@ -364,7 +364,7 @@ impl Calculator {
             for samples in &samples {
                 let mut builder = LabelsBuilder::new(vec!["sample"]);
                 for sample_i in 0..samples.count() {
-                    builder.add(&[sample_i]);
+                    builder.add(&[sample_i as i32]);
                 }
                 cell_gradient_samples.push(builder.finish_assume_unique());
             }
@@ -385,7 +385,7 @@ impl Calculator {
             for samples in &samples {
                 let mut builder = LabelsBuilder::new(vec!["sample"]);
                 for sample_i in 0..samples.count() {
-                    builder.add(&[sample_i]);
+                    builder.add(&[sample_i as i32]);
                 }
                 strain_gradient_samples.push(builder.finish_assume_unique());
             }
